wip: show revert status
In CLI git, it typically shows this:
```
On branch main
Your branch is ahead of 'origin/main' by 2 commits.
  (use "git push" to publish your local commits)

You are currently reverting commit 57409cb.
  (fix conflicts and run "git revert --continue")
  (use "git revert --skip" to skip this patch)
  (use "git revert --abort" to cancel the revert operation)

Unmerged paths:
  (use "git restore --staged <file>..." to unstage)
  (use "git add/rm <file>..." as appropriate to mark resolution)
        deleted by them: new-file

no changes added to commit (use "git add" and/or "git commit -a")
```

No need to show all of this, but an indicator that we are in the
"reverting state" would be useful, so that the user can
revert --continue, revert --abort etc.